requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
PyMySQL==1.1.1
Werkzeug==3.1.3
requests==2.31.0
gunicorn==21.2.0
